requirements.txt
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
tabulate==0.9.0
instagrapi==2.0.0
